# src/api/swaggerV6.yaml
swagger: '2.0'
info:
  description: Governify Registry
  version: '6'
  title: Governify Registry
  termsOfService: www.isa.us.es
  contact:
    name: ISA Research Group
    url: www.isa.us.es
    email: isagroup.us@gmail.com
  license:
    name: MIT
    url: 'http://opensource.org/licenses/MIT'
host: registry.governify.io
basePath: /api/v6
...
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
...
  '/agreements/{agreement}':
    get:
      tags:
        - agreement-registry
      description: Get an agreement by ID
      operationId: getAgreementById
      parameters:
        - name: agreement
          in: path
          description: XXX
          required: true
          type: string
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/agreement'
        '404':
          description: Not found
          schema:
            $ref: '#/definitions/error_model'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error_model'
      x-swagger-router-controller: AgreementController
    delete:
      tags:
        - agreement-registry
      description: XXX
      operationId: deleteAgreementById
      parameters:
        - name: agreement
          in: path
          description: XXX
          required: true
          type: string
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/agreement'
        '400':
          description: Bad request
          schema:
            $ref: '#/definitions/error_model'
      x-swagger-router-controller: AgreementController
  /states:
    get:
      tags:
        - state-registry
      description: XXX
      operationId: statesGET
      responses:
        '200':
          description: Success
          schema:
            type: array
            items:
              $ref: '#/definitions/agreement_state'
        '400':
          description: Bad request
          schema:
            $ref: '#/definitions/error_model'
      x-swagger-router-controller: StateController
    delete:
      tags:
        - state-registry
      description: XXX
      operationId: statesDELETE
      responses:
        '200':
          description: Success
        '400':
          description: Bad request
          schema:
            $ref: '#/definitions/error_model'
      x-swagger-router-controller: StateController
  '/states/{agreement}/filtered':
    get:
      tags:
        - state-registry
      description: XXX
      operationId: statesFilter
      parameters:
        - name: agreement
          in: path
          description: XXX
          required: true
          type: string
      responses:
        '200':
          description: Success
          schema:
            type: array
            items:
              $ref: '#/definitions/agreement_state'
        '400':
          description: Bad request
          schema:
            $ref: '#/definitions/error_model'
      x-swagger-router-controller: StateController
  '/states/{agreement}':
    get:
      tags:
        - state-registry
      description: XXX
      operationId: getAgreementById
      parameters:
        - name: agreement
          in: path
          description: XXX
          required: true
          type: string
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/agreement_state'
        '400':
          description: Bad request
          schema:
            $ref: '#/definitions/error_model'
      x-swagger-router-controller: StateController
    delete:
      tags:
        - state-registry
      description: XXX
      operationId: deleteAgreementById
      parameters:
        - name: agreement
          in: path
          description: XXX
          required: true
          type: string
      responses:
        '200':
          description: Success
        '400':
          description: Bad request
          schema:
            $ref: '#/definitions/error_model'
      x-swagger-router-controller: StateController
...
  '/states/{agreement}/guarantees':
    get:
      tags:
        - state-registry
      description: XXX
      operationId: getAllGuarantees
      parameters:
        - name: agreement
          in: path
          description: XXX
          required: true
          type: string
      responses:
        '200':
          description: Success
          schema:
            type: array
            items:
              $ref: '#/definitions/guarantee_state'
        '400':
          description: Bad request
          schema:
            $ref: '#/definitions/error_model'
      x-swagger-router-controller: StateController
...
definitions:
...
  agreement:
    type: object
    # properties:
    #   id:
    #     type: string
    #   version:
    #     type: string
    #   type:
    #     type: string
    #   context:
    #     $ref: '#/definitions/context'
    #   terms:
    #     $ref: '#/definitions/terms'
  context:
    type: object
    properties:
      provider:
        type: string
      consumer:
        type: string
      validity:
        $ref: '#/definitions/validity'
      definitions:
        $ref: '#/definitions/definitions'
      infrastructure:
        $ref: '#/definitions/infrastructure'
  validity:
    type: object
    required:
      - initial
      - timeZone
    properties:
      initial:
        type: string
      timeZone:
        type: string
      end:
        type: string
  infrastructure:
    type: object
  definitions:
    type: object
    properties:
      schemas:
        $ref: '#/definitions/schema'
      scopes:
        $ref: '#/definitions/scopes'
      logs:
        $ref: '#/definitions/log'
  schema:
    type: object
  scopes:
    type: object
  scope:
    type: object
  log:
    type: object
  terms:
    type: object
    properties:
      guarantees:
        type: array
        items:
          $ref: '#/definitions/guarantee'
      configurations:
        type: object
        additionalProperties:
          $ref: '#/definitions/configuration'
      metrics:
        $ref: '#/definitions/metric'
  configuration:
    type: object
    properties:
      scope:
        type: object
      of:
        type: array
        items:
          $ref: '#/definitions/scopedConfiguration'
  scopedConfiguration:
    type: object
    properties:
      scope:
        type: object
      value:
        type: string
  metric:
    type: object
...
  guarantee:
    type: object
    properties:
      id:
        type: string
      scope:
        $ref: '#/definitions/guarantee_scope'
      of:
        type: array
        items:
          $ref: '#/definitions/scoped_guarantee'
  guarantee_scope:
    type: object
  scoped_guarantee:
    type: object
    properties:
      scope:
        $ref: '#/definitions/guarantee_scope'
      objective:
        type: string
      with:
        $ref: '#/definitions/with'
      window:
        $ref: '#/definitions/window'
      evidences:
        type: array
        items:
          $ref: '#/definitions/evidence'
  with:
    type: object
  window:
    type: object
    required:
      - type
      - period
    properties:
      initial:
        type: string
      end:
        type: string
      type:
        type: string
      period:
        type: string
  evidence:
    type: object
...
  agreement_state:
    type: object
...
  guarantee_state:
    type: object
    properties:
      guarantee:
        type: string
      scope:
        $ref: '#/definitions/scope'
      logs:
        $ref: '#/definitions/log'
      period:
        $ref: '#/definitions/period'
      evidences:
        type: array
        items:
          $ref: '#/definitions/evidence_state'
      fulfilled:
        type: boolean
  period:
    type: object
    properties:
      from:
        type: string
      to:
        type: string
...
  evidence_state:
    type: object
    properties:
      evidence:
        type: boolean
      issue_init:
        type: string
      issue_end:
        type: string
      issue_duration:
        type: string
...
  error_model:
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
...